create more stable camera movement.

+++ EmnityDX/Objects/LevelData/PlayingStateLevels/TestLevel.cs
@@ -115,7 +115,7 @@ namespace EmnityDX.Objects.LevelData.PlayingStateLevels
                         PositionComponents[entity.Id] = pos;
                     }
 
-                    camera.Position = new Vector2(PositionComponents[entity.Id].Pos.X + SpriteComponents[entity.Id].SpritePath.Bounds.Center.X, PositionComponents[entity.Id].Pos.Y + SpriteComponents[entity.Id].SpritePath.Bounds.Center.Y);
+                    camera.Position = new Vector2((int)PositionComponents[entity.Id].Pos.X + SpriteComponents[entity.Id].SpritePath.Bounds.Center.X, (int)PositionComponents[entity.Id].Pos.Y + SpriteComponents[entity.Id].SpritePath.Bounds.Center.Y);
                 }
                 else if ((entity.ComponentFlags & ComponentMasks.MOVEMENT) == ComponentMasks.MOVEMENT)
                 {
